#2021 - Refactor curation code
- Remove SuggestionBuilder
- Remove code which was used for correction/automation in WebAnno
