+++ inception-ui-curation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/curation/page/CurationPage.java
@@ -17,12 +17,20 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.ui.curation.page;
 
+import static de.tudarmstadt.ukp.clarin.webanno.api.CasUpgradeMode.AUTO_CAS_UPGRADE;
 import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.CURATION_USER;
 import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.AnnotatorStateUtils.updateDocumentTimestampAfterWrite;
 import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.AnnotatorStateUtils.verifyAndUpdateDocumentTimestamp;
 import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.page.AnnotationPageBase.PAGE_PARAM_DOCUMENT;
 import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.paging.FocusPosition.CENTERED;
 import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.paging.FocusPosition.TOP;
+import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil.getAddr;
+import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil.getFirstSentence;
+import static de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasAccessMode.SHARED_READ_ONLY_ACCESS;
+import static de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasAccessMode.UNMANAGED_ACCESS;
+import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.doDiffSingle;
+import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.getDiffAdapters;
+import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.LinkCompareBehavior.LINK_ROLE_AS_LABEL;
 import static de.tudarmstadt.ukp.clarin.webanno.model.PermissionLevel.CURATOR;
 import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.CURATION_FINISHED;
 import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentStateTransition.ANNOTATION_IN_PROGRESS_TO_CURATION_IN_PROGRESS;
@@ -30,6 +38,10 @@ import static de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaBehavior.en
 import static de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaBehavior.visibleWhen;
 import static de.tudarmstadt.ukp.clarin.webanno.ui.core.page.ProjectPageBase.NS_PROJECT;
 import static de.tudarmstadt.ukp.clarin.webanno.ui.core.page.ProjectPageBase.PAGE_PARAM_PROJECT;
+import static de.tudarmstadt.ukp.clarin.webanno.ui.curation.component.model.SentenceState.AGREE;
+import static de.tudarmstadt.ukp.clarin.webanno.ui.curation.component.model.SentenceState.DISAGREE;
+import static org.apache.uima.fit.util.CasUtil.getType;
+import static org.apache.uima.fit.util.CasUtil.selectCovered;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -39,9 +51,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.uima.UIMAException;
 import org.apache.uima.cas.CAS;
+import org.apache.uima.cas.Type;
+import org.apache.uima.cas.text.AnnotationFS;
+import org.apache.uima.fit.util.CasUtil;
 import org.apache.wicket.AttributeModifier;
 import org.apache.wicket.ajax.AjaxEventBehavior;
 import org.apache.wicket.ajax.AjaxRequestTarget;
@@ -70,6 +86,7 @@ import org.wicketstuff.event.annotation.OnEvent;
 import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.clarin.webanno.api.DocumentService;
 import de.tudarmstadt.ukp.clarin.webanno.api.ProjectService;
+import de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.AnnotationEditorBase;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.actionbar.ActionBar;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.event.AnnotationEvent;
@@ -80,11 +97,18 @@ import de.tudarmstadt.ukp.clarin.webanno.api.annotation.page.AnnotationPageBase;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.paging.SentenceOrientedPagingStrategy;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.rendering.event.RenderAnnotationsEvent;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.rendering.event.SelectionChangedEvent;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil;
 import de.tudarmstadt.ukp.clarin.webanno.brat.annotation.BratAnnotationEditor;
 import de.tudarmstadt.ukp.clarin.webanno.constraints.ConstraintsService;
+import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.Configuration;
+import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.ConfigurationSet;
+import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.DiffResult;
+import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.api.DiffAdapter;
+import de.tudarmstadt.ukp.clarin.webanno.curation.casmerge.CasMerge;
 import de.tudarmstadt.ukp.clarin.webanno.curation.storage.CurationDocumentService;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationDocument;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationDocumentState;
+import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 import de.tudarmstadt.ukp.clarin.webanno.model.Mode;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocument;
@@ -92,6 +116,7 @@ import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState;
 import de.tudarmstadt.ukp.clarin.webanno.security.UserDao;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.Role;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.User;
+import de.tudarmstadt.ukp.clarin.webanno.support.StopWatch;
 import de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaBehavior;
 import de.tudarmstadt.ukp.clarin.webanno.support.wicket.DecoratedObject;
 import de.tudarmstadt.ukp.clarin.webanno.support.wicket.WicketUtil;
@@ -102,7 +127,8 @@ import de.tudarmstadt.ukp.clarin.webanno.ui.curation.component.model.AnnotationS
 import de.tudarmstadt.ukp.clarin.webanno.ui.curation.component.model.AnnotatorSegment;
 import de.tudarmstadt.ukp.clarin.webanno.ui.curation.component.model.SentenceIndex;
 import de.tudarmstadt.ukp.clarin.webanno.ui.curation.component.model.SentenceInfo;
-import de.tudarmstadt.ukp.clarin.webanno.ui.curation.component.model.SuggestionBuilder;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token;
 
 /**
  * This is the main class for the curation page. It contains an interface which displays differences
@@ -217,13 +243,13 @@ public class CurationPage
         curationView = new SentenceInfo();
 
         List<AnnotatorSegment> segments = new LinkedList<>();
-        AnnotatorSegment userAnnotationSegments = new AnnotatorSegment();
+        AnnotatorSegment annotatorSegment = new AnnotatorSegment();
 
         if (getModelObject() != null) {
-            userAnnotationSegments
+            annotatorSegment
                     .setSelectionByUsernameAndAddress(annotationSelectionByUsernameAndAddress);
-            userAnnotationSegments.setAnnotatorState(getModelObject());
-            segments.add(userAnnotationSegments);
+            annotatorSegment.setAnnotatorState(getModelObject());
+            segments.add(annotatorSegment);
         }
 
         // update source list model only first time.
@@ -506,9 +532,7 @@ public class CurationPage
 
             currentprojectId = state.getProject().getId();
 
-            SuggestionBuilder builder = new SuggestionBuilder(documentService,
-                    curationDocumentService, annotationService);
-            curationContainer = builder.buildCurationContainer(state);
+            curationContainer = buildCurationContainer(state);
             detailPanel.reset(aTarget);
             init(aTarget, curationContainer);
 
@@ -549,11 +573,10 @@ public class CurationPage
 
         AnnotationDocument randomAnnotationDocument = finishedAnnotationDocuments.get(0);
 
-        SuggestionBuilder cb = new SuggestionBuilder(documentService, curationDocumentService,
-                annotationService);
-        Map<String, CAS> casses = cb.readAllCasesSharedNoUpgrade(finishedAnnotationDocuments);
-        CAS mergeCas = cb.getMergeCas(state, state.getDocument(), casses, randomAnnotationDocument,
+        Map<String, CAS> casses = readAllCasesSharedNoUpgrade(finishedAnnotationDocuments);
+        CAS mergeCas = getMergeCas(state, state.getDocument(), casses, randomAnnotationDocument,
                 true, aMergeIncompleteAnnotations, aForceRecreateCas);
+
         return mergeCas;
     }
 
@@ -817,5 +840,257 @@ public class CurationPage
             allSourceDocuments.add(dsd);
         }
         return allSourceDocuments;
+    }
+
+    public SentenceIndex buildCurationContainer(AnnotatorState aState)
+        throws UIMAException, ClassNotFoundException, IOException, AnnotationException
+    {
+        // get annotation documents
+        List<AnnotationDocument> finishedAnnotationDocuments = documentService
+                .listFinishedAnnotationDocuments(aState.getDocument());
+
+        Map<String, CAS> casses = readAllCasesSharedNoUpgrade(finishedAnnotationDocuments);
+        CAS mergeCas = getMergeCas(aState, aState.getDocument(), casses, null, false, false, false);
+
+        int diffWindowStart = getFirstSentence(mergeCas).getBegin();
+        int diffWindowEnd = mergeCas.getDocumentText().length();
+
+        Map<Integer, Integer> segmentBeginEnd = new HashMap<>();
+        Map<Integer, Integer> segmentNumber = new HashMap<>();
+        Map<String, Map<Integer, Integer>> segmentAdress = new HashMap<>();
+        updateSegment(aState, segmentBeginEnd, segmentNumber, segmentAdress, mergeCas,
+                CURATION_USER, diffWindowStart, diffWindowEnd);
+
+        segmentAdress.put(CURATION_USER, new HashMap<>());
+        Type sentenceType = getType(mergeCas, Sentence.class);
+        for (AnnotationFS s : selectCovered(mergeCas, sentenceType, diffWindowStart,
+                diffWindowEnd)) {
+            segmentAdress.get(CURATION_USER).put(s.getBegin(), getAddr(s));
+        }
+
+        List<DiffAdapter> adapters = getDiffAdapters(annotationService,
+                aState.getAnnotationLayers());
+
+        long diffStart = System.currentTimeMillis();
+        LOG.debug("Calculating differences...");
+        int count = 0;
+        SentenceIndex curationContainer = new SentenceIndex();
+        for (Integer begin : segmentBeginEnd.keySet()) {
+            Integer end = segmentBeginEnd.get(begin);
+
+            count++;
+            if (count % 100 == 0) {
+                LOG.debug("Processing differences: {} of {} sentences...", count,
+                        segmentBeginEnd.size());
+            }
+
+            DiffResult diff = doDiffSingle(adapters, LINK_ROLE_AS_LABEL, casses, begin, end)
+                    .toResult();
+
+            SentenceInfo curationSegment = new SentenceInfo(begin, end, segmentNumber.get(begin));
+
+            if (diff.hasDifferences() || !diff.getIncompleteConfigurationSets().isEmpty()) {
+                // Is this confSet a diff due to stacked annotations (with same configuration)?
+                boolean stackedDiff = false;
+
+                stackedDiffSet: for (ConfigurationSet d : diff.getDifferingConfigurationSets()
+                        .values()) {
+                    for (Configuration c : d.getConfigurations()) {
+                        if (c.getCasGroupIds().size() != d.getCasGroupIds().size()) {
+                            stackedDiff = true;
+                            break stackedDiffSet;
+                        }
+                    }
+                }
+
+                if (stackedDiff) {
+                    curationSegment.setSentenceState(DISAGREE);
+                }
+                else if (!diff.getIncompleteConfigurationSets().isEmpty()) {
+                    curationSegment.setSentenceState(DISAGREE);
+                }
+                else {
+                    curationSegment.setSentenceState(AGREE);
+                }
+            }
+            else {
+                curationSegment.setSentenceState(AGREE);
+            }
+
+            for (String username : segmentAdress.keySet()) {
+                curationSegment.getSentenceAddressByUserIndex().put(username,
+                        segmentAdress.get(username).get(begin));
+            }
+
+            curationContainer.addSentenceInfo(curationSegment);
+        }
+        LOG.debug("Difference calculation completed in {}ms",
+                (System.currentTimeMillis() - diffStart));
+
+        return curationContainer;
+    }
+
+    public Map<String, CAS> readAllCasesSharedNoUpgrade(List<AnnotationDocument> aDocuments)
+        throws IOException
+    {
+        Map<String, CAS> casses = new HashMap<>();
+        for (AnnotationDocument annDoc : aDocuments) {
+            String username = annDoc.getUser();
+            CAS cas = documentService.readAnnotationCas(annDoc.getDocument(), username,
+                    AUTO_CAS_UPGRADE, SHARED_READ_ONLY_ACCESS);
+            casses.put(username, cas);
+        }
+        return casses;
+    }
+
+    /**
+     * Fetches the CAS that the user will be able to edit. In AUTOMATION/CORRECTION mode, this is
+     * the CAS for the CORRECTION_USER and in CURATION mode it is the CAS for the CURATION user.
+     *
+     * @param aState
+     *            the model.
+     * @param aDocument
+     *            the source document.
+     * @param aCasses
+     *            the CASes.
+     * @param aTemplate
+     *            an annotation document which is used as a template for the new merge CAS.
+     * @return the CAS.
+     * @throws UIMAException
+     *             hum?
+     * @throws ClassNotFoundException
+     *             hum?
+     * @throws IOException
+     *             if an I/O error occurs.
+     * @throws AnnotationException
+     *             hum?
+     */
+    public CAS getMergeCas(AnnotatorState aState, SourceDocument aDocument,
+            Map<String, CAS> aCasses, AnnotationDocument aTemplate, boolean aUpgrade,
+            boolean aMergeIncompleteAnnotations, boolean aForceRecreateCas)
+        throws UIMAException, ClassNotFoundException, IOException, AnnotationException
+    {
+        if (aForceRecreateCas) {
+            return initializeMergeCas(aState, aCasses, aTemplate, aMergeIncompleteAnnotations);
+        }
+
+        if (!curationDocumentService.existsCurationCas(aDocument)) {
+            return initializeMergeCas(aState, aCasses, aTemplate, aMergeIncompleteAnnotations);
+        }
+
+        CAS mergeCas = curationDocumentService.readCurationCas(aDocument);
+        if (aUpgrade) {
+            curationDocumentService.upgradeCurationCas(mergeCas, aDocument);
+            curationDocumentService.writeCurationCas(mergeCas, aDocument, true);
+            updateDocumentTimestampAfterWrite(aState,
+                    curationDocumentService.getCurationCasTimestamp(aState.getDocument()));
+        }
+
+        return mergeCas;
+    }
+
+    public CAS initializeMergeCas(AnnotatorState aState, Map<String, CAS> aCasses,
+            AnnotationDocument aTemplate, boolean aMergeIncompleteAnnotations)
+        throws ClassNotFoundException, UIMAException, IOException, AnnotationException
+    {
+        CAS mergeCas = createCurationCas(aState, aTemplate, aCasses, aState.getAnnotationLayers(),
+                aMergeIncompleteAnnotations);
+        updateDocumentTimestampAfterWrite(aState,
+                curationDocumentService.getCurationCasTimestamp(aState.getDocument()));
+
+        return mergeCas;
+    }
+
+    /**
+     * Puts CASes into a list and get a random annotation document that will be used as a base for
+     * the diff.
+     */
+    private void updateSegment(AnnotatorState aBratAnnotatorModel,
+            Map<Integer, Integer> aIdxSentenceBeginEnd,
+            Map<Integer, Integer> aIdxSentenceBeginNumber,
+            Map<String, Map<Integer, Integer>> aSegmentAdress, CAS aCas, String aUsername,
+            int aWindowStart, int aWindowEnd)
+    {
+        // Get the number of the first sentence - instead of fetching the number over and over
+        // we can just increment this one.
+        int sentenceNumber = WebAnnoCasUtil.getSentenceNumber(aCas, aWindowStart);
+
+        aSegmentAdress.put(aUsername, new HashMap<>());
+        Type sentenceType = CasUtil.getType(aCas, Sentence.class);
+        for (AnnotationFS sentence : selectCovered(aCas, sentenceType, aWindowStart, aWindowEnd)) {
+            aIdxSentenceBeginEnd.put(sentence.getBegin(), sentence.getEnd());
+            aIdxSentenceBeginNumber.put(sentence.getBegin(), sentenceNumber);
+            aSegmentAdress.get(aUsername).put(sentence.getBegin(), getAddr(sentence));
+            sentenceNumber += 1;
+        }
+    }
+
+    @Deprecated
+    public static List<Type> getEntryTypes(CAS aMergeCas, List<AnnotationLayer> aLayers,
+            AnnotationSchemaService aAnnotationService)
+    {
+        List<Type> entryTypes = new LinkedList<>();
+
+        for (AnnotationLayer layer : aLayers) {
+            if (layer.getName().equals(Token.class.getName())) {
+                continue;
+            }
+            if (layer.getType().equals(WebAnnoConst.CHAIN_TYPE)) {
+                continue;
+            }
+            entryTypes.add(aAnnotationService.getAdapter(layer).getAnnotationType(aMergeCas));
+        }
+        return entryTypes;
+    }
+
+    /**
+     * For the first time a curation page is opened, create a MergeCas that contains only agreeing
+     * annotations Using the CAS of the curator user.
+     *
+     * @param aState
+     *            the annotator state
+     * @param aRandomAnnotationDocument
+     *            an annotation document.
+     * @param aCasses
+     *            the CASes
+     * @param aAnnotationLayers
+     *            the layers.
+     * @return the CAS.
+     * @throws IOException
+     *             if an I/O error occurs.
+     */
+    private CAS createCurationCas(AnnotatorState aState,
+            AnnotationDocument aRandomAnnotationDocument, Map<String, CAS> aCasses,
+            List<AnnotationLayer> aAnnotationLayers, boolean aMergeIncompleteAnnotations)
+        throws IOException, UIMAException, AnnotationException
+    {
+        Validate.notNull(aState, "State must be specified");
+        Validate.notNull(aRandomAnnotationDocument, "Annotation document must be specified");
+
+        // We need a modifiable copy of some annotation document which we can use to initialize
+        // the curation CAS. This is an exceptional case where BYPASS is the correct choice
+        CAS mergeCas = documentService.readAnnotationCas(aRandomAnnotationDocument,
+                UNMANAGED_ACCESS);
+
+        List<DiffAdapter> adapters = getDiffAdapters(annotationService,
+                aState.getAnnotationLayers());
+
+        DiffResult diff;
+        try (StopWatch watch = new StopWatch(LOG, "CasDiff")) {
+            diff = doDiffSingle(adapters, LINK_ROLE_AS_LABEL, aCasses, 0,
+                    mergeCas.getDocumentText().length()).toResult();
+        }
+
+        try (StopWatch watch = new StopWatch(LOG, "CasMerge")) {
+            CasMerge casMerge = new CasMerge(annotationService);
+            casMerge.setMergeIncompleteAnnotations(aMergeIncompleteAnnotations);
+            casMerge.reMergeCas(diff, aState.getDocument(), aState.getUser().getUsername(),
+                    mergeCas, aCasses);
+        }
+
+        curationDocumentService.writeCurationCas(mergeCas, aRandomAnnotationDocument.getDocument(),
+                false);
+
+        return mergeCas;
     }
 }
